STILL BROKEN: edits to core and os-specific code are largely done and now I'm just fixing the driver code

+++ libvmi/driver/interface.c
@@ -22,6 +22,10 @@ struct driver_instance{
     void (*set_name_ptr)(vmi_instance_t, char *);
     status_t (*get_memsize_ptr)(vmi_instance_t, unsigned long *);
     status_t (*get_vcpureg_ptr)(vmi_instance_t, reg_t *, registers_t, unsigned long);
+    unsigned long (*pfn_to_mfn_ptr)(vmi_instance_t, unsigned long);
+    void *(*map_page_ptr)(vmi_instance_t, int, unsigned long);
+    void *(*map_pages_ptr)(vmi_instance_t, int, unsigned long *, unsigned long);
+    int (*is_pv_ptr)(vmi_instance_t);
 };
 typedef struct driver_instance * driver_instance_t;
 
@@ -29,42 +33,82 @@ driver_instance_t instance = NULL;
 xen_instance_t xeninst;
 file_instance_t fileinst;
 
+void driver_xen_setup (vmi_instance_t vmi)
+{
+    vmi->driver = &xeninst;
+    instance->init_ptr = &xen_init;
+    instance->destroy_ptr = &xen_destroy;
+    instance->get_id_ptr = &xen_get_domainid;
+    instance->set_id_ptr = &xen_set_domainid;
+    instance->get_name_ptr = &xen_get_domainname;
+    //TODO add set_name_ptr
+    instance->get_memsize_ptr = &xen_get_memsize;
+    instance->get_vcpureg_ptr = &xen_get_vcpureg;
+    instance->pfn_to_mfn_ptr = &xen_pfn_to_mfn;
+    instance->map_page_ptr = &xen_map_page;
+    instance->map_pages_ptr = &xen_map_pages;
+    instance->is_pv_ptr = &xen_is_pv;
+}
+
+void driver_kvm_setup (vmi_instance_t vmi)
+{
+    //TODO set vmi->driver
+    //TODO add init_ptr
+    //TODO add destroy_ptr
+    //TODO add get_id_ptr
+    //TODO add set_id_ptr
+    //TODO add get_name_ptr
+    //TODO add set_name_ptr
+    //TODO add get_memsize_ptr
+    //TODO add get_vcpureg_ptr
+    //TODO add pfn_to_mfn_ptr
+    //TODO add map_page_ptr
+    //TODO add map_pages_ptr
+    //TODO add is_pv_ptr
+}
+
+void driver_file_setup (vmi_instance_t vmi)
+{
+    vmi->driver = &fileinst;
+    instance->init_ptr = &file_init;
+    //TODO add destroy_ptr
+    //TODO add get_id_ptr
+    //TODO add set_id_ptr
+    //TODO add get_name_ptr
+    instance->set_name_ptr = &file_set_name;
+    instance->get_memsize_ptr = &file_get_memsize;
+    instance->get_vcpureg_ptr = &file_get_vcpureg;
+    instance->pfn_to_mfn_ptr = &file_pfn_to_mfn;
+    instance->map_page_ptr = &file_map_page;
+    //TODO add map_pages_ptr
+    instance->is_pv_ptr = &file_is_pv;
+}
+
 driver_instance_t driver_get_instance (vmi_instance_t vmi)
 {
-    if (NULL != instance){
-        return instance;
-    }
-    else{
+    if (NULL == instance){
         /* allocate memory for the function pointers */
         instance = (driver_instance_t) safe_malloc(sizeof(struct driver_instance));
 
         /* assign the function pointers */
         if (VMI_MODE_XEN == vmi->mode){
-            instance->driver = &xeninst;
-            instance->init_ptr = &xen_init;
-            instance->destroy_ptr = &xen_destroy;
-            instance->get_id_ptr = &xen_get_domainid;
-            instance->set_id_ptr = &xen_set_domainid;
-            instance->get_name_ptr = &xen_get_domainname;
-            //TODO add set_name_ptr
-            instance->get_memsize_ptr = &xen_get_memsize;
-            instance->get_vcpureg_ptr = &xen_get_vcpureg;
+            driver_xen_setup(vmi);
         }
         else if (VMI_MODE_KVM == vmi->mode){
+            driver_kvm_setup(vmi);
         }
         else if (VMI_MODE_FILE == vmi->mode){
-            instance->driver = &fileinst;
-            instance->init_ptr = &file_init;
-            instance->set_name_ptr = &file_set_name;
-            instance->get_memsize_ptr = &file_get_memsize;
+            driver_file_setup(vmi);
         }
+
     }
+    return instance;
 }
 
 status_t driver_init (vmi_instance_t vmi)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
+    if (NULL != ptrs && NULL != ptrs->init_ptr){
         return ptrs->init_ptr(vmi);
     }
     else{
@@ -75,7 +119,7 @@ status_t driver_init (vmi_instance_t vmi)
 void driver_destroy (vmi_instance_t vmi)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
+    if (NULL != ptrs && NULL != ptrs->destroy_ptr){
         return ptrs->destroy_ptr(vmi);
     }
     else{
@@ -86,7 +130,7 @@ void driver_destroy (vmi_instance_t vmi)
 unsigned long driver_get_id (vmi_instance_t vmi)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
+    if (NULL != ptrs && NULL != ptrs->get_id_ptr){
         return ptrs->get_id_ptr(vmi);
     }
     else{
@@ -94,10 +138,10 @@ unsigned long driver_get_id (vmi_instance_t vmi)
     }
 }
 
-void driver_set_id (vmi_instance_t vmi, unsigned long id);
+void driver_set_id (vmi_instance_t vmi, unsigned long id)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
+    if (NULL != ptrs && NULL != ptrs->set_id_ptr){
         return ptrs->set_id_ptr(vmi, id);
     }
     else{
@@ -108,18 +152,18 @@ void driver_set_id (vmi_instance_t vmi, unsigned long id);
 status_t driver_get_name (vmi_instance_t vmi, char **name)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
-        return ptrs->get_vmname_ptr(vmi, name);
+    if (NULL != ptrs && NULL != ptrs->get_name_ptr){
+        return ptrs->get_name_ptr(vmi, name);
     }
     else{
         return VMI_FAILURE;
     }
 }
 
-void driver_set_name (vmi_instance_t vmi, char *name);
+void driver_set_name (vmi_instance_t vmi, char *name)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
+    if (NULL != ptrs && NULL != ptrs->set_name_ptr){
         return ptrs->set_name_ptr(vmi, name);
     }
     else{
@@ -130,7 +174,7 @@ void driver_set_name (vmi_instance_t vmi, char *name);
 status_t driver_get_memsize (vmi_instance_t vmi, unsigned long *size)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
+    if (NULL != ptrs && NULL != ptrs->get_memsize_ptr){
         return ptrs->get_memsize_ptr(vmi, size);
     }
     else{
@@ -141,7 +185,7 @@ status_t driver_get_memsize (vmi_instance_t vmi, unsigned long *size)
 status_t driver_get_vcpureg (vmi_instance_t vmi, reg_t *value, registers_t reg, unsigned long vcpu)
 {
     driver_instance_t ptrs = driver_get_instance(vmi);
-    if (NULL != ptrs){
+    if (NULL != ptrs && NULL != ptrs->get_vcpureg_ptr){
         return ptrs->get_vcpureg_ptr(vmi, value, reg, vcpu);
     }
     else{
@@ -149,15 +193,46 @@ status_t driver_get_vcpureg (vmi_instance_t vmi, reg_t *value, registers_t reg, 
     }
 }
 
+unsigned long driver_pfn_to_mfn (vmi_instance_t vmi, unsigned long pfn)
+{
+    driver_instance_t ptrs = driver_get_instance(vmi);
+    if (NULL != ptrs && NULL != ptrs->pfn_to_mfn_ptr){
+        return ptrs->pfn_to_mfn_ptr(vmi, pfn);
+    }
+    else{
+        return 0;
+    }
+}
 
+void *driver_map_page (vmi_instance_t vmi, int prot, unsigned long page)
+{
+    driver_instance_t ptrs = driver_get_instance(vmi);
+    if (NULL != ptrs && NULL != ptrs->map_page_ptr){
+        return ptrs->map_page_ptr(vmi, prot, page);
+    }
+    else{
+        return NULL;
+    }
+}
 
+void *driver_map_pages (vmi_instance_t vmi, int prot, unsigned long *pages, unsigned long num_pages)
+{
+    driver_instance_t ptrs = driver_get_instance(vmi);
+    if (NULL != ptrs && NULL != ptrs->map_pages_ptr){
+        return ptrs->map_pages_ptr(vmi, prot, pages, num_pages);
+    }
+    else{
+        return NULL;
+    }
+}
 
-
-// lookup domain name from id // xen, used in read_config_file from core.c
-// lookup domain id from name // xen, used in init_name functions
-
-// pfn -> mfn conversion
-
-// map a single page
-
-// map multiple (possible non-contiguous) pages // used in xa_access_user_va_range
+int driver_is_pv (vmi_instance_t vmi)
+{
+    driver_instance_t ptrs = driver_get_instance(vmi);
+    if (NULL != ptrs && NULL != ptrs->is_pv_ptr){
+        return ptrs->is_pv_ptr(vmi);
+    }
+    else{
+        return 0;
+    }
+}
